Refactoring Services for the Vacuums
